Add material and purchase request
