--- material_request/data/sequence.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-    <data noupdate="0">
-
-        <record id="sequence_material_request" model="ir.sequence">
-            <field name="name">Purchase Request</field>
-            <field name="code">material_request.material_request</field>
-            <field name="prefix">MR/</field>
-            <field name="padding">5</field>
-        </record>        
-
-    </data>
-</odoo>
