--- material_request/security/ir.model.access.csv
@@ -1,2 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_material_request_material_request,material_request.material_request,model_material_request_material_request,,1,0,0,0
