+++ material_requisition/models/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from . import ir_request
-from . import ir_request_line
+# from . import ir_request_line
 from . import stock_picking
 from . import res_company
-from . import purchase_requisition
+# from . import purchase_requisition
